+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [#3167](https://github.com/bbatsov/rubocop/issues/3167): Add new `Style/VariableNumber` cop. ([@sooyang][])
 * Add new style `no_mixed_keys` to `Style/HashSyntax` to only check for hashes with mixed keys. ([@daviddavis][])
 * Allow including multiple configuration files from a single gem. ([@tjwallace][])
+* Add check for `persisted?` method call when using a create method in `Rails/SaveBang`. ([@QuinnHarris][])
 
 ### Bug fixes